app/requirements.txt
django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt
psycopg2-binary==2.9.9